refactor numerated using deps
