+++ lazy-pages/src/pages.rs
@@ -18,7 +18,11 @@
 
 //! Module for pages which size can be different for different runtime versions.
 
-use numerated::{Bound, Interval, IntervalIterator, Numerated, OptionBound};
+use numerated::{
+    interval::Interval,
+    iterators::IntervalIterator,
+    numerated::{Bound, Numerated, OptionBound},
+};
 use std::{cmp::Ordering, marker::PhantomData, mem, num::NonZeroU32};
 
 pub trait SizeNumber: Copy + Ord + Eq {
